@@ -7,7 +7,7 @@ members = [
 name = "tmux-backup"
 version = "0.2.0"
 edition = "2021"
-rust-version = "1.59.0"
+rust-version = "1.60.0"
 description = "A backup & restore solution for Tmux sessions."
 readme = "README.md"
 
