@@ -51,7 +51,7 @@ impl Metadata {
     /// Query Tmux and return a new `Metadata`.
     pub async fn new() -> Result<Self> {
         let version = FORMAT_VERSION.to_string();
-        let client = tmux::client::current_client().await?;
+        let client = tmux::client::current().await?;
         let sessions = tmux::session::available_sessions().await?;
         let windows = tmux::window::available_windows().await?;
         let panes = tmux::pane::available_panes().await?;
